+++ cmake/nfxGraphicsSources.cmake
@@ -4,6 +4,7 @@
 
 set(NFX_GRAPHICS_SOURCES
     src/gl/material/BlinnPhong.cpp
+    src/gl/material/Unlit.cpp
     src/gl/material/Material.cpp
     src/gl/mesh/Mesh.cpp
     src/gl/mesh/Primitive.cpp